refactor: assignment.ts・shuffle.ts を lib/firestore/assignment に移設し syncStatus 通知を追加 (#546)
assignment データ層移行シリーズ(#542)の最終段(4/4)。

- assignment.ts/shuffle.ts を lib/firestore/assignment/ へ移設
- 購読(subscribeLatestAssignmentDay/subscribeShuffleEvent)を createSyncedSubscription で包み購読失敗を reportSyncError へ
- 書き込み・読み取りを runWriteWithSync で包み保存失敗を reportSaveError へ
- updateAssignmentDay の console.error のみの自前 catch を撤去(通知は mutateAssignmentDay 内で実施)
- fetchRecentShuffleHistory は orderBy フォールバックを保つため try/catch 全体を runWriteWithSync で包む
- B&B 大域最適化(PR #494)のロジックには非干渉、データ層の移送のみ
- 呼び出し元6ファイルの import を @/lib/firestore/assignment へ張り替え
- 旧 app/assignment/lib/firebase/ ディレクトリ全体を削除(シム含む)
- 未使用化したバレル再エクスポート(toMillisSafe 等4件)を削除し Dead Code Check を green に
