+++ lib/firestore/assignment/assignment.ts
@@ -1,3 +1,7 @@
+// assignment 担当データ本体（issue #546）
+// 旧 app/assignment/lib/firebase/assignment.ts から移設。
+// 購読は createSyncedSubscription 経由で失敗時に reportSyncError、
+// 書き込み・読み取りは runWriteWithSync 経由で失敗時に reportSaveError につなぐ。
 import {
   doc,
   getDocs,
@@ -9,14 +13,12 @@ import {
   limit,
   runTransaction,
 } from 'firebase/firestore';
+import type { QuerySnapshot } from 'firebase/firestore';
 import { db } from '@/lib/firebase';
 import { AssignmentDay, Assignment } from '@/types';
-import {
-  getAssignmentDaysCollection,
-  normalizeAssignmentsForDate,
-  sortAssignmentsStable,
-  areAssignmentsEqual,
-} from './helpers';
+import { getAssignmentDaysCollection } from './references';
+import { normalizeAssignmentsForDate, sortAssignmentsStable, areAssignmentsEqual } from './helpers';
+import { createSyncedSubscription, runWriteWithSync } from './sync';
 
 export const getServerTodayDate = async (timeZone: string = 'Asia/Tokyo'): Promise<string> => {
   return new Intl.DateTimeFormat('en-CA', { timeZone }).format(new Date());
@@ -30,39 +32,37 @@ export const mutateAssignmentDay = async (
   const assignmentDaysCol = getAssignmentDaysCollection(userId);
   const docRef = doc(assignmentDaysCol, date);
 
-  return runTransaction(db, async (tx) => {
-    const snap = await tx.get(docRef);
-    const existingData = snap.exists() ? (snap.data() as AssignmentDay) : undefined;
-    const currentRaw = existingData?.assignments ?? [];
-    const normalizedCurrent = sortAssignmentsStable(normalizeAssignmentsForDate(currentRaw, date));
+  return runWriteWithSync(() =>
+    runTransaction(db, async (tx) => {
+      const snap = await tx.get(docRef);
+      const existingData = snap.exists() ? (snap.data() as AssignmentDay) : undefined;
+      const currentRaw = existingData?.assignments ?? [];
+      const normalizedCurrent = sortAssignmentsStable(normalizeAssignmentsForDate(currentRaw, date));
 
-    const proposed = sortAssignmentsStable(normalizeAssignmentsForDate(updater(currentRaw), date));
+      const proposed = sortAssignmentsStable(normalizeAssignmentsForDate(updater(currentRaw), date));
 
-    const shouldCreateDoc = !snap.exists();
-    const hasDifference = shouldCreateDoc || !areAssignmentsEqual(normalizedCurrent, proposed);
+      const shouldCreateDoc = !snap.exists();
+      const hasDifference = shouldCreateDoc || !areAssignmentsEqual(normalizedCurrent, proposed);
 
-    if (!hasDifference) {
-      return { assignments: normalizedCurrent, changed: false };
-    }
+      if (!hasDifference) {
+        return { assignments: normalizedCurrent, changed: false };
+      }
 
-    tx.set(docRef, {
-      date,
-      assignments: proposed,
-      updatedAt: serverTimestamp(),
-      createdAt: existingData?.createdAt ?? serverTimestamp(),
-    });
+      tx.set(docRef, {
+        date,
+        assignments: proposed,
+        updatedAt: serverTimestamp(),
+        createdAt: existingData?.createdAt ?? serverTimestamp(),
+      });
 
-    return { assignments: proposed, changed: true };
-  });
+      return { assignments: proposed, changed: true };
+    })
+  );
 };
 
 export const updateAssignmentDay = async (userId: string, date: string, assignments: Assignment[]) => {
-  try {
-    await mutateAssignmentDay(userId, date, () => assignments);
-  } catch (error) {
-    console.error('Failed to update assignment day:', error);
-    throw error;
-  }
+  // 失敗時の通知・再 throw は mutateAssignmentDay 内の runWriteWithSync が担う
+  await mutateAssignmentDay(userId, date, () => assignments);
 };
 
 export const subscribeLatestAssignmentDay = (
@@ -74,8 +74,8 @@ export const subscribeLatestAssignmentDay = (
   const latestQuery = query(assignmentDaysCol, orderBy('updatedAt', 'desc'), limit(1));
   let initializing = false;
 
-  return onSnapshot(
-    latestQuery,
+  return createSyncedSubscription<QuerySnapshot>(
+    (onNext, onError) => onSnapshot(latestQuery, onNext, onError),
     async (snap) => {
       if (!snap.empty) {
         const docSnap = snap.docs[0];
@@ -90,15 +90,13 @@ export const subscribeLatestAssignmentDay = (
         try {
           await options.onEmpty();
         } catch (error) {
+          // onEmpty 内の実書き込みは updateAssignmentDay 側で通知済み。ここは購読コールバックを
+          // 例外で止めないための保険なので console.error に留める
           console.error('Failed to initialize first assignment day:', error);
         } finally {
           initializing = false;
         }
       }
-    },
-    (error) => {
-      console.error('Failed to subscribe latest assignment day:', error);
-      callback(null);
     }
   );
 };
@@ -108,7 +106,7 @@ export const fetchRecentAssignments = async (
   endDate: string,
   days: number
 ): Promise<AssignmentDay[]> => {
-  try {
+  return runWriteWithSync(async () => {
     const assignmentDaysCol = getAssignmentDaysCollection(userId);
     const promises = [];
     const targetDate = new Date(endDate);
@@ -130,8 +128,5 @@ export const fetchRecentAssignments = async (
     });
 
     return results;
-  } catch (error) {
-    console.error('Failed to fetch recent assignments:', error);
-    throw error;
-  }
+  });
 };
