+++ lib/firestore/assignment/index.ts
@@ -1,16 +1,26 @@
 // lib/firestore/assignment バレルエクスポート（issue #543）
 
-// コレクション参照
-export { getAssignmentDaysCollection, getShuffleEventsCollection, getShuffleHistoryCollection } from './references';
-
-// 汎用ロジック・定数
+// 担当データ本体（取得・購読・更新）
 export {
-  toMillisSafe,
-  DEFAULT_SHUFFLE_SETTINGS,
-  normalizeAssignmentsForDate,
-  sortAssignmentsStable,
-  areAssignmentsEqual,
-} from './helpers';
+  getServerTodayDate,
+  mutateAssignmentDay,
+  updateAssignmentDay,
+  subscribeLatestAssignmentDay,
+  fetchRecentAssignments,
+} from './assignment';
+
+// シャッフルイベント・履歴
+export {
+  subscribeShuffleEvent,
+  createShuffleEvent,
+  updateShuffleEventState,
+  createShuffleHistory,
+  fetchRecentShuffleHistory,
+} from './shuffle';
+
+// 汎用ロジック・定数（toMillisSafe 等は各モジュールが ./helpers から直接 import するため
+// バレルには消費者のいる定数のみ公開する。未使用 export は Dead Code Check で検出される）
+export { DEFAULT_SHUFFLE_SETTINGS } from './helpers';
 
 // 設定管理（テーブル/シャッフル/管理者/ペア除外）
 export {
